rename discover to community
