+++ e2e/src/features/community/detail-pages.feature
