+++ e2e/src/features/community/smoke.feature
