+++ e2e/src/mocks/community/types.ts
