+++ e2e/src/steps/community/smoke.steps.ts
